+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [Improved] Moved the lattitude/longitude info down below the address information for on the Site Identity & Site Creator settings
 * [Improved] Wrapped all `templates->render()` calls in try/catch blocks to more gracefully handle errors
 * [Added] Added an `Anonymize IP` toggle for Google Analytics, for compliance with some local laws
+* [Added] Added Wikipedia to the Social Media settings & `sameAs` entity information
 
 ## 1.1.41 -- 2017.02.10